user_server/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
jinja2==3.1.4
python-dotenv==1.0.1
pinecone-client==5.0.1
google-generativeai==0.8.3
langchain-text-splitters==0.3.5
pdfplumber==0.11.4
python-multipart==0.0.20
requests==2.32.3
pydantic==2.10.6
pypdf==5.1.0
pytesseract==0.3.13
pillow==10.4.0
sqlalchemy
